backend/requirements.txt
asgiref==3.11.0
Django==6.0
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
PyJWT==2.10.1
pytz==2025.2
sqlparse==0.5.5
psycopg2-binary==2.9.11
python-dotenv==1.2.1
Pillow==12.0.0
reportlab==4.4.7
whitenoise==6.11.0
gunicorn==21.2.0
requests==2.32.5
twilio==9.9.0
dj-database-url==2.1.0




